requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
pydantic
pydantic-settings
openai
pgvector
python-multipart